Window code cleanup
